refactor(async-migrations): extract async migrations engine into its own product
Moves the async-migrations framework engine out of core
`posthog/async_migrations/` into `products/async_migrations/backend/`, so
its ~1k LOC of slow ClickHouse-touching tests no longer run as part of the
core `posthog/` suite.

Scope is engine + tests only (the second PR in the core-extraction stack,
on top of the approvals extraction):

- `posthog/async_migrations/` → `products/async_migrations/backend/`
  (history preserved via git rename). The hardcoded discovery paths in
  `setup.py` (`ASYNC_MIGRATIONS_MODULE_PATH`) are repointed; all 10
  migration definitions still resolve, and DB-stored migration names
  (derived from file basenames) are unchanged.
- The `AsyncMigration` / `AsyncMigrationError` models stay in
  `posthog/models/` — no model migration. The API viewset, Celery tasks,
  instance status/settings, app-startup hook, and the
  `run_async_migrations` management command also stay in core; their
  imports are repointed to the product.
- Deliberately a package-only product (no Django `AppConfig`, not in
  `PRODUCTS_APPS`): it has no Django models or routes, and its
  `migrations/` subpackage holds ClickHouse async-migration definitions —
  registering it as a Django app would make Django try to load those as
  Django migrations, a regression that does not exist today.
- The dedicated CI job that runs the async-migration tests
  (`-m "async_migrations"`) is repointed to the new path in both the
  GitHub workflow and its Depot shadow, so the tests keep running.
- Frontend source link and the async-migrations handbook doc updated to
  the new path; `tach.toml` (core → `products.async_migrations`) and
  `products/ruff.toml` exemptions added per the existing per-product
  convention.

Validated locally: framework and all consumers import, all 10 migration
definitions are discovered under the new path, `makemigrations --check`
reports no drift (no model moved), and `tach check` passes.

Generated-By: PostHog Code
Task-Id: e7c94753-d29d-4cad-8bf5-dee4ccb40d34
