+++ posthog/management/commands/run_async_migrations.py
@@ -7,8 +7,6 @@ from django.core.management.base import BaseCommand
 import structlog
 from semantic_version.base import Version
 
-from posthog.async_migrations.runner import complete_migration, is_migration_dependency_fulfilled, start_async_migration
-from posthog.async_migrations.setup import ALL_ASYNC_MIGRATIONS, setup_async_migrations, setup_model
 from posthog.clickhouse.query_tagging import Feature, Product, tags_context
 from posthog.constants import FROZEN_POSTHOG_VERSION
 from posthog.models.async_migration import (
@@ -19,6 +17,13 @@ from posthog.models.async_migration import (
     is_async_migration_complete,
 )
 from posthog.models.instance_setting import get_instance_setting
+
+from products.async_migrations.backend.runner import (
+    complete_migration,
+    is_migration_dependency_fulfilled,
+    start_async_migration,
+)
+from products.async_migrations.backend.setup import ALL_ASYNC_MIGRATIONS, setup_async_migrations, setup_model
 
 logger = structlog.get_logger(__name__)
 logger.setLevel(logging.INFO)
