+++ posthog/tasks/async_migrations.py
@@ -1,14 +1,15 @@
 from celery import shared_task, states
 from celery.result import AsyncResult
 
-from posthog.async_migrations.runner import (
+from posthog.models.instance_setting import get_instance_setting
+
+from products.async_migrations.backend.runner import (
     run_async_migration_operations,
     run_migration_healthcheck,
     start_async_migration,
     update_migration_progress,
 )
-from posthog.async_migrations.utils import force_stop_migration, process_error, trigger_migration
-from posthog.models.instance_setting import get_instance_setting
+from products.async_migrations.backend.utils import force_stop_migration, process_error, trigger_migration
 
 
 # we're hijacking an entire worker to do this - consider:
