+++ posthog/tasks/test/test_async_migrations.py
@@ -7,9 +7,9 @@ from unittest.mock import patch
 from celery import states
 from celery.result import AsyncResult
 
-from posthog.async_migrations.examples.test_migration import Migration
-from posthog.async_migrations.runner import run_async_migration_next_op, run_async_migration_operations
-from posthog.async_migrations.test.util import create_async_migration
+from products.async_migrations.backend.examples.test_migration import Migration
+from products.async_migrations.backend.runner import run_async_migration_next_op, run_async_migration_operations
+from products.async_migrations.backend.test.util import create_async_migration
 from posthog.models.async_migration import AsyncMigration, MigrationStatus
 from posthog.models.instance_setting import set_instance_setting
 from posthog.tasks.async_migrations import check_async_migration_health
