+++ products/async_migrations/backend/status.py
@@ -1,6 +1,7 @@
 def async_migrations_ok() -> bool:
-    from posthog.async_migrations.runner import is_posthog_version_compatible
     from posthog.models.async_migration import AsyncMigration, MigrationStatus
+
+    from products.async_migrations.backend.runner import is_posthog_version_compatible
 
     for migration in AsyncMigration.objects.all():
         migration_completed_or_running = migration.status in [
