+++ products/async_migrations/backend/test/test_0007_persons_and_groups_on_events_backfill.py
@@ -4,9 +4,6 @@ from uuid import uuid4
 import pytest
 from posthog.test.base import ClickhouseTestMixin, run_clickhouse_statement_in_parallel
 
-from posthog.async_migrations.runner import start_async_migration
-from posthog.async_migrations.setup import get_async_migration_definition, setup_async_migrations
-from posthog.async_migrations.test.util import AsyncMigrationBaseTest
 from posthog.clickhouse.client import query_with_columns, sync_execute
 from posthog.models.async_migration import AsyncMigration, AsyncMigrationError, MigrationStatus
 from posthog.models.event.util import create_event
@@ -14,6 +11,10 @@ from posthog.models.group.util import create_group
 from posthog.models.person.util import create_person, create_person_distinct_id, delete_person
 from posthog.models.utils import UUIDT
 from posthog.test.persons import create_person as create_test_person
+
+from products.async_migrations.backend.runner import start_async_migration
+from products.async_migrations.backend.setup import get_async_migration_definition, setup_async_migrations
+from products.async_migrations.backend.test.util import AsyncMigrationBaseTest
 
 pytestmark = pytest.mark.async_migrations
