+++ products/async_migrations/backend/test/test_definition.py
@@ -5,21 +5,22 @@ from django.test import SimpleTestCase
 
 from infi.clickhouse_orm.utils import import_submodules
 
-from posthog.async_migrations.definition import AsyncMigrationDefinition, AsyncMigrationOperation
-from posthog.async_migrations.setup import (
+from posthog.models.async_migration import AsyncMigration
+from posthog.version_requirement import ServiceVersionRequirement
+
+from products.async_migrations.backend.definition import AsyncMigrationDefinition, AsyncMigrationOperation
+from products.async_migrations.backend.setup import (
     ASYNC_MIGRATIONS_EXAMPLE_MODULE_PATH,
     get_async_migration_definition,
     setup_async_migrations,
 )
-from posthog.models.async_migration import AsyncMigration
-from posthog.version_requirement import ServiceVersionRequirement
 
 pytestmark = pytest.mark.async_migrations
 
 
 class TestAsyncMigrationDefinition(SimpleTestCase):
     def test_get_async_migration_definition(self):
-        from posthog.async_migrations.examples.example import example_fn, example_rollback_fn
+        from products.async_migrations.backend.examples.example import example_fn, example_rollback_fn
 
         modules = import_submodules(ASYNC_MIGRATIONS_EXAMPLE_MODULE_PATH)
         example_migration = modules["example"].Migration("example")
