+++ products/async_migrations/package.json
@@ -1,0 +1,7 @@
+{
+    "name": "@posthog/products-async-migrations",
+    "scripts": {
+        "backend:test": "pytest -c ../../pytest.ini --rootdir ../.. backend/test -v --tb=short",
+        "backend:lint": "ruff check --config ../../pyproject.toml backend && ruff format --config ../../pyproject.toml --check backend"
+    }
+}
