+++ products/async_migrations/product.yaml
@@ -1,0 +1,3 @@
+name: Async migrations
+owners:
+    - clickhouse
