+++ products/ruff.toml
@@ -22,6 +22,7 @@ extend-select = ["ANN001", "ANN201", "ANN202", "ANN204"]
 "alerts/backend/**" = ["ANN"]
 "annotations/backend/**" = ["ANN"]
 "approvals/backend/**" = ["ANN"]
+"async_migrations/backend/**" = ["ANN"]
 "batch_exports/backend/**" = ["ANN"]
 "cdp/backend/**" = ["ANN"]
 "cohorts/backend/**" = ["ANN"]
@@ -62,6 +63,7 @@ extend-select = ["ANN001", "ANN201", "ANN202", "ANN204"]
 "analytics_platform/**" = ["PLC0415"]
 "annotations/**" = ["PLC0415"]
 "approvals/**" = ["PLC0415"]
+"async_migrations/**" = ["PLC0415"]
 "batch_exports/**" = ["PLC0415"]
 "business_knowledge/**" = ["PLC0415"]
 "cdp/**" = ["PLC0415"]
